Fix cr problems
