+++ be/src/olap/rowset_builder.h
@@ -15,33 +15,31 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#ifndef DORIS_BE_SRC_ROWSET_ROWSET_H
-#define DORIS_BE_SRC_ROWSET_ROWSET_H
+#ifndef DORIS_BE_SRC_ROWSET_ROWSET_WRITER_H
+#define DORIS_BE_SRC_ROWSET_ROWSET_WRITER_H
 
-#include "olap/olap_define.h"
-#include "rowset/rowset_reader.h"
-#include "rowset/rowset_writer.h"
-
-#include <memory>
+#include "rowset/rowset.h"
+#include "olap/new_status.h"
+#include "olap/schema.h"
+#include "olap/row_block.h"
 
 namespace doris {
 
-class Rowset {
+class RowsetBuilder {
 public:
-    // 初始化Rowset
-    OLAPStatus init(RowsetMeta rowset_meta, const DataDir* dir) = 0;
+	NewStatus init(std::string rowset_id, const std::string& rowset_path_prefix, Schema* schema) = 0;
 
-    std::unique_ptr<RowsetReader> create_reader() = 0;
+    // add a row block to rowset
+	NewStatus add_row_block(RowBlock* row_block) = 0;
 
-    OLAPStatus copy(RowsetWriter* dest_rowset_writer) = 0;
+    // this is a temp api
+    // it is used to get rewritten path for writing rowset data
+	NewStatus generate_written_path(const std::string& src_path, std::string* dest_path) = 0;
 
-    OLAPStatus delete() = 0;
-
-private:
-    DataDir* _dir;	
-	RowsetMeta _rowset_meta;
+    // get a rowset
+	NewStatus build(Rowset* rowset) = 0;
 };
 
 }
 
-#endif // DORIS_BE_SRC_ROWSET_ROWSET_H
+#endif // DORIS_BE_SRC_ROWSET_ROWSET_WRITER_H
