+++ gensrc/proto/olap_file.proto
@@ -71,22 +71,20 @@ enum RowsetState {
 
 message RowsetMetaPb {
     required int64 rowset_id = 1;
-    // RowsetMeta update version
-    required int64 version = 2;
-    optional int64 tablet_id = 3;
-    optional int32 tablet_schema_hash = 4;
-    optional RowsetType rowset_type = 5;
-    optional RowsetState rowset_state = 6;
+    optional int64 tablet_id = 2;
+    optional int32 tablet_schema_hash = 3;
+    optional RowsetType rowset_type = 4;
+    optional RowsetState rowset_state = 5;
     // Rowset data version range
-    optional int32 start_version = 7;
-    optional int32 end_version = 8;
-    optional int32 row_number = 9;
-    optional int64 total_disk_size = 10;
-    optional int64 data_disk_size = 11;
-    optional int64 index_disk_size = 12;
+    optional int32 start_version = 6;
+    optional int32 end_version = 7;
+    optional int32 row_number = 8;
+    optional int64 total_disk_size = 9;
+    optional int64 data_disk_size = 10;
+    optional int64 index_disk_size = 11;
     // column min/max/null flag statistic info
-    repeated ColumnPruning column_statistics = 13;
-    optional DeleteConditionMessage delete_condition = 14;
+    repeated ColumnPruning column_statistics = 12;
+    optional DeleteConditionMessage delete_predicate = 13;
     // spare field id 15-49 for future use
     optional bytes extra_properties = 50;
 }
